+++ lib/dogapi/facade.rb
@@ -12,24 +12,20 @@ module Dogapi
   # documentation, here[https://github.com/DataDog/dogapi/wiki].
   class Client
     # Create a new Client optionally specifying a default host and device
-    def initialize(api_key, application_key = nil, host = nil, device = nil, silent = true, timeout = nil)
+    def initialize(api_key, application_key=nil, host=nil, device=nil, silent=true, timeout=nil, endpoint=nil)
       raise 'Please provide an API key to submit your data' if api_key.nil?
       @api_key = api_key
-
       @application_key = application_key
-
-      @datadog_host = Dogapi.find_datadog_host
-
+      @datadog_host = endpoint || Dogapi.find_datadog_host()
       @host = host || Dogapi.find_localhost
-
       @device = device
+      @legacy_event_svc = Dogapi::EventService.new(@datadog_host)
 
       %w(
-        metric event tag comment search dash alert user snapshot embed
-        screenboard monitor service_check
+        metric event tag comment search dash alert user snapshot embed screenboard
+        monitor service_check
       ).each do |service_type|
         service_class = service_type.split('_').map(&:capitalize).join
-
         instance_variable_set(
           "@#{service_type}_svc",
           Dogapi.const_get("Dogapi::V1::#{service_class}Service").new(
@@ -37,8 +33,6 @@ module Dogapi
           )
         )
       end
-
-      @legacy_event_svc = Dogapi::EventService.new(@datadog_host)
     end
 
     #
@@ -55,8 +49,9 @@ module Dogapi
     #
     # options[:type] = "counter" to specify a counter metric
     # options[:tags] = ["tag1", "tag2"] to tag the point
-    def emit_point(metric, value, options = {})
-      options = { timestamp: Time.now }.merge(options)
+    def emit_point(metric, value, options= {})
+      defaults = { :timestamp => Time.now }
+      options = defaults.merge(options)
 
       emit_points(
         metric,
@@ -76,12 +71,11 @@ module Dogapi
     #
     # options[:type] = "counter" to specify a counter metric
     # options[:tags] = ["tag1", "tag2"] to tag the point
-    def emit_points(metric, points, options = {})
+    def emit_points(metric, points, options= {})
       scope = override_scope options
 
       points.each do |p|
         raise 'Not a Time' unless p[0].is_a? Time
-
         p[0] = p[0].to_i
         p[1] = p[1].to_f # TODO: stupid to_f will never raise an exception
       end
@@ -118,7 +112,7 @@ module Dogapi
     # Optional arguments:
     #  :host        => String
     #  :device      => String
-    def emit_event(event, options = {})
+    def emit_event(event, options= {})
       scope = override_scope options
 
       @event_svc.post(event, scope)
@@ -131,6 +125,13 @@ module Dogapi
       @event_svc.get(id)
     end
 
+    # Delete an event
+    #
+    # +id+ of the event to delete
+    def delete_event(id)
+      @event_svc.delete(id)
+    end
+
     # Get an optionally filtered event stream
     #
     # +start+ is a Time object for when to start the stream
@@ -141,7 +142,7 @@ module Dogapi
     #   :priority   => "normal" or "low"
     #   :sources    => String, see https://github.com/DataDog/dogapi/wiki/Event for a current list of sources
     #   :tags       => Array of Strings
-    def stream(start, stop, options = {})
+    def stream(start, stop, options= {})
       @event_svc.stream(start, stop, options)
     end
 
@@ -149,7 +150,8 @@ module Dogapi
     # The functionality will be removed in a later release.
     def start_event(event, options = {})
       deprecation_warning('start_event', 'emit_event')
-      options = { source_type:  nil }.merge(options)
+      defaults = { :source_type => nil }
+      options = defaults.merge(options)
 
       scope = override_scope options
 
@@ -163,12 +165,12 @@ module Dogapi
     #
 
     # Post a comment
-    def comment(message, options = {})
+    def comment(message, options= {})
       @comment_svc.comment(message, options)
     end
 
     # Post a comment
-    def update_comment(comment_id, options = {})
+    def update_comment(comment_id, options= {})
       @comment_svc.update_comment(comment_id, options)
     end
 
@@ -190,14 +192,14 @@ module Dogapi
     #
 
     # Get all tags and their associated hosts at your org
-    def all_tags(source = nil)
+    def all_tags(source=nil)
       @tag_svc.get_all(source)
     end
 
     # Get all tags for the given host
     #
     # +host_id+ can be the host's numeric id or string name
-    def host_tags(host_id, source = nil, by_source = false)
+    def host_tags(host_id, source=nil, by_source=false)
       @tag_svc.get(host_id, source, by_source)
     end
 
@@ -206,7 +208,7 @@ module Dogapi
     # +host_id+ can be the host's numeric id or string name
     #
     # +tags+ is and Array of Strings
-    def add_tags(host_id, tags, source = nil)
+    def add_tags(host_id, tags, source=nil)
       @tag_svc.add(host_id, tags, source)
     end
 
@@ -215,7 +217,7 @@ module Dogapi
     # +host_id+ can be the host's numeric id or string name
     #
     # +tags+ is and Array of Strings
-    def update_tags(host_id, tags, source = nil)
+    def update_tags(host_id, tags, source=nil)
       @tag_svc.update(host_id, tags, source)
     end
 
@@ -228,7 +230,7 @@ module Dogapi
     # Remove all tags from the given host
     #
     # +host_id+ can be the host's numeric id or string name
-    def detach_tags(host_id, source = nil)
+    def detach_tags(host_id, source=nil)
       @tag_svc.detach(host_id, source)
     end
 
@@ -237,12 +239,12 @@ module Dogapi
     #
 
     # Create a dashboard.
-    def create_dashboard(title, description, graphs, template_variables = nil)
+    def create_dashboard(title, description, graphs, template_variables=nil)
       @dash_service.create_dashboard(title, description, graphs, template_variables)
     end
 
     # Update a dashboard.
-    def update_dashboard(dash_id, title, description, graphs, template_variables = nil)
+    def update_dashboard(dash_id, title, description, graphs, template_variables=nil)
       @dash_service.update_dashboard(dash_id, title, description, graphs, template_variables)
     end
 
@@ -265,11 +267,11 @@ module Dogapi
     # ALERTS
     #
 
-    def alert(query, options = {})
+    def alert(query, options= {})
       @alert_svc.alert(query, options)
     end
 
-    def update_alert(alert_id, query, options = {})
+    def update_alert(alert_id, query, options= {})
       @alert_svc.update_alert(alert_id, query, options)
     end
 
@@ -294,11 +296,11 @@ module Dogapi
     end
 
     # User invite
-    def invite(emails, options = {})
+    def invite(emails, options= {})
       @user_svc.invite(emails, options)
     end
 
-    def create_user(description = {})
+    def create_user(description= {})
       @user_svc.create_user(description)
     end
 
@@ -310,7 +312,7 @@ module Dogapi
       @user_svc.get_user(handle)
     end
 
-    def update_user(handle, description = {})
+    def update_user(handle, description= {})
       @user_svc.update_user(handle, description)
     end
 
@@ -319,7 +321,7 @@ module Dogapi
     end
 
     # Graph snapshot
-    def graph_snapshot(metric_query, start_ts, end_ts, event_query = nil)
+    def graph_snapshot(metric_query, start_ts, end_ts, event_query=nil)
       @snapshot_svc.snapshot(metric_query, start_ts, end_ts, event_query)
     end
 
@@ -382,15 +384,15 @@ module Dogapi
     # MONITORS
     #
 
-    def monitor(type, query, options = {})
+    def monitor(type, query, options= {})
       @monitor_svc.monitor(type, query, options)
     end
 
-    def update_monitor(monitor_id, query, options = {})
+    def update_monitor(monitor_id, query, options= {})
       @monitor_svc.update_monitor(monitor_id, query, options)
     end
 
-    def get_monitor(monitor_id, options = {})
+    def get_monitor(monitor_id, options= {})
       @monitor_svc.get_monitor(monitor_id, options)
     end
 
@@ -398,7 +400,7 @@ module Dogapi
       @monitor_svc.delete_monitor(monitor_id)
     end
 
-    def get_all_monitors(options = {})
+    def get_all_monitors(options= {})
       @monitor_svc.get_all_monitors(options)
     end
 
@@ -410,11 +412,11 @@ module Dogapi
       @monitor_svc.unmute_monitors
     end
 
-    def mute_monitor(monitor_id, options = {})
+    def mute_monitor(monitor_id, options= {})
       @monitor_svc.mute_monitor(monitor_id, options)
     end
 
-    def unmute_monitor(monitor_id, options = {})
+    def unmute_monitor(monitor_id, options= {})
       @monitor_svc.unmute_monitor(monitor_id, options)
     end
 
@@ -422,11 +424,11 @@ module Dogapi
     # MONITOR DOWNTIME
     #
 
-    def schedule_downtime(scope, options = {})
+    def schedule_downtime(scope, options= {})
       @monitor_svc.schedule_downtime(scope, options)
     end
 
-    def update_downtime(downtime_id, options = {})
+    def update_downtime(downtime_id, options= {})
       @monitor_svc.update_downtime(downtime_id, options)
     end
 
@@ -438,7 +440,7 @@ module Dogapi
       @monitor_svc.cancel_downtime(downtime_id)
     end
 
-    def get_all_downtimes(options = {})
+    def get_all_downtimes(options= {})
       @monitor_svc.get_all_downtimes(options)
     end
 
@@ -446,7 +448,7 @@ module Dogapi
     # HOST MUTING
     #
 
-    def mute_host(hostname, options = {})
+    def mute_host(hostname, options= {})
       @monitor_svc.mute_host(hostname, options)
     end
 
@@ -458,7 +460,7 @@ module Dogapi
     # SERVICE CHECKS
     #
 
-    def service_check(check, host, status, options = {})
+    def service_check(check, host, status, options= {})
       @service_check_svc.service_check(check, host, status, options)
     end